Create skeleton
